+++ Source/EffectChain.cpp
@@ -161,7 +161,7 @@ void EffectChain::Process(double time)
    for (int ch = 0; ch < GetBuffer()->NumActiveChannels(); ++ch)
    {
       float* buffer = GetBuffer()->GetChannel(ch);
-      float volSq = mVolume * mVolume;
+      double volSq = mVolume * mVolume;
       Mult(buffer, volSq, bufferSize);
       Add(target->GetBuffer()->GetChannel(ch), buffer, bufferSize);
       GetVizBuffer()->WriteChunk(buffer, bufferSize, ch);
