+++ Source/ModuleSaveDataPanel.cpp
@@ -99,8 +99,8 @@ void ModuleSaveDataPanel::ReloadSaveData()
       maxWidth = MAX(maxWidth, GetStringWidth(*iter));
 
    mAlignmentX = 10 + maxWidth;
-   float x = mAlignmentX;
-   float y = 5 + kItemSpacing;
+   double x = mAlignmentX;
+   double y = 5 + kItemSpacing;
 
    mNameEntry = new TextEntry(this, "", x, y, 27, mSaveModule->NameMutable());
    mNameEntry->SetNoHover(true);
